fix: memory_optimize is less aggressive
tsutils.memory_optimize was very aggressive in terms of finding the
smallest dtype.  Felt too aggressive and changed to fundamentally using
df.convert_dtypes(). Affected many of the tests which had expected the
old way.

Can now have {} in a docstring that doesn't have a variable name in
between them.

Started to use pytest to expect warnings and deprecations.

Improved tsutils.read_iso_ts() to better accommodate bytes - perhaps
replace completely the capture function in tests/capture.py.

+++ tests/test_dropna.py
@@ -12,10 +12,10 @@ class TestRead(TestCase):
             index_col=[0],
             parse_dates=True,
             skipinitialspace=True,
-        ).astype("float64")
+        ).astype("Float64")
         self.read_direct.index.name = "Datetime"
 
     def test_read_direct_dropna(self):
         """Test read dropna for single column - daily."""
-        out = tstoolbox.read("tests/data_missing.csv", dropna="all")
+        out = tstoolbox.read("tests/data_missing.csv", dropna="all").astype("Float64")
         assert_frame_equal(out, self.read_direct)
